Keep question follow tests behavioral
